+++ src/main/java/be/ugent/idlab/knows/dataio/export/Export.java
@@ -3,7 +3,7 @@ package be.ugent.idlab.knows.dataio.export;
 import java.io.OutputStream;
 
 /**
- * Interface representing outputting to a data source.
+ * Interface representing outputting to a data sink.
  */
 public interface Export {
     OutputStream getOutputStream() throws Exception;